@@ -2,7 +2,6 @@
     "name": "twig/extensions",
     "description": "Common additional features for Twig that do not directly belong in core",
     "keywords": ["i18n","text"],
-    "homepage": "http://twig.sensiolabs.org/doc/extensions/index.html",
     "type": "library",
     "license": "MIT",
     "authors": [
@@ -12,11 +11,11 @@
         }
     ],
     "require": {
-        "twig/twig": "~1.27|~2.0"
+        "twig/twig": "^1.27|^2.0"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "~3.3@dev",
-        "symfony/translation": "~2.3|~3.0"
+        "symfony/phpunit-bridge": "^3.4",
+        "symfony/translation": "^2.7|^3.4|^4.2"
     },
     "suggest": {
         "symfony/translation": "Allow the time_diff output to be translated"
