@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-class Twig_Tests_Extension_IntlTest extends PHPUnit_Framework_TestCase
+class Twig_Tests_Extension_IntlTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @requires extension intl
@@ -22,5 +22,17 @@ class Twig_Tests_Extension_IntlTest extends PHPUnit_Framework_TestCase
         $env = $this->getMockBuilder('Twig_Environment')->disableOriginalConstructor()->getMock();
         $date = twig_localized_date_filter($env, new DateTime('2015-01-01T00:00:00', new DateTimeZone('UTC')), 'short', 'long', 'en', '+01:00');
         $this->assertEquals('1/1/15, 1:00:00 AM GMT+1', $date);
+    }
+
+    /**
+     * @requires extension intl
+     * @requires PHP 5.5
+     */
+    public function testLocalizedDateFilterWithDateTimeZoneZ()
+    {
+        class_exists('Twig_Extensions_Extension_Intl');
+        $env = $this->getMockBuilder('Twig_Environment')->disableOriginalConstructor()->getMock();
+        $date = twig_localized_date_filter($env, new DateTime('2017-11-19T00:00:00Z'), 'short', 'long', 'fr', 'Z');
+        $this->assertEquals('19/11/2017 00:00:00 UTC', $date);
     }
 }
